% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_r.R
\name{cor_data_set_m_sd}
\alias{cor_data_set_m_sd}
\title{Change the mean and standard deviation of each column in a data frame to specified values}
\usage{
cor_data_set_m_sd(cor.data, means.in, sds.in)
}
\arguments{
\item{cor.data}{Data frame}

\item{means.in}{A vector of desired column means. The number means must correspond to the number of columns.}

\item{sds.in}{A vector of desired column standard deviations. The number standard deviations must correspond to the number of columns.}
}
\value{
Data frame with the desired means and standard deviations for the columns
}
\description{
Change the mean and standard deviation of each column in a data frame to specified values
}
\examples{
N <- 600
M <- matrix(c(1.00, 0.60, 0.30, 0.30,
            0.60, 1.00, 0.00, 0.60,
            0.30, 0.00, 1.00, 0.00,
            0.30, 0.60, 0.00, 1.00), nrow=4, ncol=4)
my.data <- get_cor_data_from_matrix(M,N)
desired.means <- c(3.5,4,4.5,5)
desired.sds <- c(1.1,1.2,1.3,1.4)
my.data <- cor_data_set_m_sd(my.data, desired.means,desired.sds)
library(apaTables)
apa.cor.table(my.data)
}

