% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-example1.R
\docType{data}
\name{example1}
\alias{example1}
\title{Example 1 Data (n=100)}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{SelfEsteem}{Self-Esteem scores }
  \item{Height}{Heights of participants}
}}
\usage{
data(example1)
}
\description{
Example 1 Data (n=100)
}
\keyword{datasets}

