% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-example2.R
\docType{data}
\name{example2}
\alias{example2}
\title{Example 2 Data (n=75)}
\format{A data frame with 75 rows and 2 variables:
\describe{
  \item{SelfEsteem}{Self-Esteem scores }
  \item{Height}{Heights of participants}
}}
\usage{
data(example2)
}
\description{
Example 2 Data (n=75)
}
\keyword{datasets}

