% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{get_cor_data}
\alias{get_cor_data}
\title{Calculate x,y data to create a specified correlation}
\usage{
get_cor_data(r, n)
}
\arguments{
\item{r}{Desired correlation}

\item{n}{Sample size}
}
\value{
Matrix with sample data
}
\description{
Calculate x,y data to create a specified correlation
}
\examples{
my.data <- get_cor_data(r=.25, n=100)
}

