% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{get_cor_samples}
\alias{get_cor_samples}
\title{Calculate a number of sample correlations based on a specified population correlation}
\usage{
get_cor_samples(rho = NA, n = NA, n.min = NA, n.max = NA,
  number.of.samples = 10, number.of.decimals = 2)
}
\arguments{
\item{rho}{Population correlation}

\item{n}{Sample size for all samples If you use n, do not use n.min or n.max.}

\item{n.min}{Minimum sample size}

\item{n.max}{Maximum sample size}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}
}
\value{
Data frame with sample correlations
}
\description{
Calculate a number of sample correlations based on a specified population correlation
}
\examples{
get_cor_samples(rho=.35,n=100)
my.samples <- get_cor_samples(rho=.35,n=100)
my.samples <- get_cor_samples(rho=.35,n.min=50,n.max=150,number.of.samples=15)
}

