% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{get_d_samples}
\alias{get_d_samples}
\title{Calculate a number of sample d-values (unbiased) based on two populations}
\usage{
get_d_samples(pop1.data = NA, pop2.data = NA, n = NA, n.min = NA,
  n.max = NA, number.of.samples = 10, number.of.decimals = 2)
}
\arguments{
\item{pop1.data}{A vector of values defining population 1}

\item{pop2.data}{A vector of values defining population 2}

\item{n}{Cell size for both cells for all samples. If you use n, do not use n.min or n.max.}

\item{n.min}{Minimum cell size across samples. Cell sizes will be equal in a single sample.}

\item{n.max}{Maximum cell size across samples. Cell sizes will be equal in a single sample.}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}
}
\value{
Data frame with sample d-values
}
\description{
Calculate a number of sample d-values (unbiased) based on two populations
}
\examples{
pop.1 <- make_population(pop.mean=100, pop.sd=15)
pop.2 <- make_population_offset_by_d(pop.data=pop.1,d=1.0)
my.samples <- get_d_samples(pop1.data=pop.1,pop2.data=pop.2,n.min=50,n.max=150,number.of.samples=15)
}

