% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{get_d_samples_infinite}
\alias{get_d_samples_infinite}
\title{Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.}
\usage{
get_d_samples_infinite(pop.d = NA, n = NA, n.min = NA, n.max = NA,
  number.of.samples = 10, number.of.decimals = 2)
}
\arguments{
\item{pop.d}{Population d-value}

\item{n}{Cell size for both cells for all samples. If you use n, do not use n.min or n.max.}

\item{n.min}{Minimum cell size across samples. Cell sizes will be equal in a single sample.}

\item{n.max}{Maximum cell size across samples. Cell sizes will be equal in a single sample.}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}
}
\value{
Data frame with sample d-values
}
\description{
Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.
}
\examples{
get_d_samples_infinite(pop.d=.35,n=100)
my.samples <- get_d_samples_infinite(pop.d=.35,n=100)
my.samples <- get_d_samples_infinite(pop.d=.35,n.min=50,n.max=150,number.of.samples=15)
}

