% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packageDetails.R
\docType{package}
\name{learnSampling}
\alias{learnSampling}
\alias{learnSampling-package}
\title{Create sampling demonstrations}
\description{
A variety of routines to help teach sampling. A few key routines are highlighted below.
\itemize{
  \item Correlations - Create correlation sampling demonstrations with \code{\link{get_cor_samples} and \link{sort_samples_by_r} }.
  \item Standardized mean differences (i.e., \emph{d}-value). Create d-value sampling demonstrations with \code{\link{make_population}, \link{make_population_offset_by_d}, \link{get_d_samples}, and \link{sort_samples_by_d} }.
 }
\tabular{ll}{
Package: \tab learnSampling\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-10-13\cr
License: \tab Unlimited\cr
}
}
\author{
\tabular{ll}{
Author: \tab David J. Stanley \email{dstanley@uoguelph.ca}\cr
Maintainer: \tab David J. Stanley \email{dstanley@uoguelph.ca}
}
}

