% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_pop_cor}
\alias{make_pop_cor}
\title{Create simulated correlation population}
\usage{
make_pop_cor(n = 1e+06, rho = 0.3, mu = c(0, 0), sd = c(1, 1))
}
\arguments{
\item{n}{Population size}

\item{rho}{Population correlation}

\item{mu}{Population mean values}

\item{sd}{Population sd values}
}
\value{
Data frame with desired properties
}
\description{
Create simulated correlation population
}
