% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_pop_d}
\alias{make_pop_d}
\title{Create simulated d-value population}
\usage{
make_pop_d(n = 1e+06, d = 0.3, mu = c(0.8, 0), sd = c(1, 1))
}
\arguments{
\item{n}{Population size (size for each of the two populations)}

\item{d}{Population d-value}

\item{mu}{Population mean values (for both populations)}

\item{sd}{Population sd values (for both populations)}
}
\value{
Data frame with desired properties
}
\description{
Create simulated d-value population
}
