% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{make_population}
\alias{make_population}
\title{Create a population with a specified mean and standard deviation}
\usage{
make_population(pop.mean = 100, pop.sd = 15, N = 10000)
}
\arguments{
\item{pop.mean}{Desired population mean}

\item{pop.sd}{d Desired population standard deviation}

\item{N}{Desired number of cases in population}
}
\value{
Vector values defining the population
}
\description{
Create a population with a specified mean and standard deviation
}
\examples{
pop.1 <- make_population(pop.mean=100, pop.sd=15)
}

