% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{make_population_offset_by_d}
\alias{make_population_offset_by_d}
\title{Calculate a second population the differs by d from an input population}
\usage{
make_population_offset_by_d(pop.data, d)
}
\arguments{
\item{pop.data}{A vector of values defining population 1}

\item{d}{The amount by which population 2 should differ from population 1}
}
\value{
Vector values for population 2
}
\description{
Calculate a second population the differs by d from an input population
}
\examples{
# Create two populations that differ by d = 1.0
pop.1 <- make_population(pop.mean=100, pop.sd=15)
pop.2 <- make_population_offset_by_d(pop.data=pop.1,d=1.0)
}

