% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mvrnorm_cor}
\alias{mvrnorm_cor}
\title{Create simulated data}
\usage{
mvrnorm_cor(n = 1, mu, sd, cormatrix, empirical = FALSE)
}
\arguments{
\item{n}{Population d-value}

\item{mu}{Population d-value}

\item{sd}{Population d-value}

\item{cormatrix}{Population d-value}

\item{empirical}{Population d-value}
}
\value{
Data frame with desired properties
}
\description{
Create simulated data
}
