% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simulate_measurement_error}
\alias{simulate_measurement_error}
\title{Create simulated cmeasurement error}
\usage{
simulate_measurement_error(df, rxx, ryy)
}
\arguments{
\item{df}{data frame with simulation results}

\item{rxx}{vector with range of rxx values}

\item{ryy}{vector with range of ryy values}
}
\value{
Data frame with desired properties
}
\description{
Create simulated cmeasurement error
}
