% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{sort_samples_by_d}
\alias{sort_samples_by_d}
\title{Sort data frame based on column d}
\usage{
sort_samples_by_d(df.in)
}
\arguments{
\item{df.in}{Data frame to sort}
}
\value{
Data frame that has been sorted by column d
}
\description{
Sort data frame based on column d
}
\examples{
my.samples <- get_d_samples(pop.d=.8,n.min=50,n.max=150,number.of.samples=10)
my.samples.sorted <- sort_samples_by_d(my.samples)
}

