% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_r.R
\name{sort_samples_by_r}
\alias{sort_samples_by_r}
\title{Sort data frame based on column r}
\usage{
sort_samples_by_r(df.in)
}
\arguments{
\item{df.in}{Data frame to sort}
}
\value{
Data frame that has been sorted by column r
}
\description{
Sort data frame based on column r
}
\examples{
my.samples <- get_cor_samples(rho=.35,n.min=50,n.max=150,number.of.samples=10)
my.samples.sorted <- sort_samples_by_r(my.samples)
}

