% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{get_d_samples_from_population_data}
\alias{get_d_samples_from_population_data}
\title{Calculate a number of sample correlations based on a specified population correlation}
\usage{
get_d_samples_from_population_data(pop1 = NULL, pop2 = NULL, cell.n,
  number.of.samples = 10, number.of.decimals = 2, var.equal = TRUE,
  alternative = "two.sided")
}
\arguments{
\item{pop1}{Population data using first column of data frame}

\item{pop2}{Population data using first column of data frame}

\item{cell.n}{Sample size for all samples (range of sample size if two specified)}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}

\item{var.equal}{TRUE/FALSE indicate type of t-test to use}

\item{alternative}{indicates type of alternative hypothesis (e.g., "two.sided") for t.test}
}
\value{
Data frame with sample correlations
}
\description{
Calculate a number of sample correlations based on a specified population correlation
}
