% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunction_means.R
\name{get_M_samples}
\alias{get_M_samples}
\title{Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.}
\usage{
get_M_samples(pop.data = NULL, data.column.name = NULL, pop.M = NA,
  pop.SD = NA, n, number.of.samples = 10, number.of.decimals = 2)
}
\arguments{
\item{pop.data}{Data to use for popuilation (optional)}

\item{data.column.name}{Column name from data to use for sampling (optional)}

\item{pop.M}{Population mean}

\item{pop.SD}{Population SD}

\item{n}{Cell size for both cells for all samples. If you use n, do not use n.min or n.max.}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}
}
\value{
Data frame with sample d-values
}
\description{
Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.
}
\examples{
get_M_samples(pop.M = 100, pop.SD = 15, n = 100)
}
