% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDetails.R
\docType{package}
\name{learnSampling}
\alias{learnSampling}
\alias{learnSampling-package}
\title{Create sampling demonstrations}
\description{
A variety of routines to help teach sampling. A few key routines are highlighted below.
\tabular{ll}{
Package: \tab learnSampling\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2019-01-29\cr
License: \tab Unlimited\cr
}
}
\author{
\tabular{ll}{
Author: \tab David J. Stanley \email{dstanley@uoguelph.ca}\cr
Maintainer: \tab David J. Stanley \email{dstanley@uoguelph.ca}
}
}
