% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mvrnorm_cor}
\alias{mvrnorm_cor}
\title{Create simulated correlatio data}
\usage{
mvrnorm_cor(n = 1, mu, sd, cormatrix, empirical = FALSE)
}
\arguments{
\item{n}{Sample size}

\item{mu}{Sample mean values}

\item{sd}{Sample sd values}

\item{cormatrix}{Sample correlation matrix}
}
\value{
Data frame with desired properties
}
\description{
Create simulated correlatio data
}
