% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplingFunctions_d.R
\name{sort_samples_by_d}
\alias{sort_samples_by_d}
\title{Sort data frame based on column d}
\usage{
sort_samples_by_d(df.in)
}
\arguments{
\item{df.in}{Data frame to sort}
}
\value{
Data frame that has been sorted by column d
}
\description{
Sort data frame based on column d
}
\examples{
pop.1 <- make_population(pop.mean=100, pop.sd=15)
pop.2 <- make_population_offset_by_d(pop.data=pop.1,d=1.0)
my.samples <- get_d_samples(pop1.data=pop.1,pop2.data=pop.2,n.min=50,n.max=150,number.of.samples=15)
my.samples.sorted <- sort_samples_by_d(my.samples)
}

